+++ backend/config/releases.exs
@@ -14,7 +14,8 @@ config :backend, Backend.Repo,
   pool_size: String.to_integer(System.get_env("POOL_SIZE") || "10"),
   ssl: ssl
 
-# show_sensitive_data_on_connection_error: true
+config :backend, Backend.Elasticsearch.Cluster,
+  url: System.get_env("ELASTICSEARCH_URL") || "http://localhost:9200"
 
 port = String.to_integer(System.get_env("PORT") || "4000")