+++ backend/lib/backend/crawler/crawlers/mastodon.ex
@@ -32,7 +32,6 @@ defmodule Backend.Crawler.Crawlers.Mastodon do
   end
 
   @impl ApiCrawler
-  # sobelow_skip ["DOS.StringToAtom"]
   def crawl(domain) do
     instance = Jason.decode!(get!("https://#{domain}/api/v1/instance").body)
 
@@ -48,7 +47,7 @@ defmodule Backend.Crawler.Crawlers.Mastodon do
     else
       Map.merge(
         Map.take(instance["stats"], ["user_count"])
-        |> Map.new(fn {k, v} -> {String.to_atom(k), v} end),
+        |> convert_keys_to_atoms(),
         %{
           peers: [],
           interactions: %{},
@@ -63,7 +62,6 @@ defmodule Backend.Crawler.Crawlers.Mastodon do
   end
 
   @spec crawl_large_instance(String.t(), any()) :: ApiCrawler.t()
-  # sobelow_skip ["DOS.StringToAtom"]
   defp crawl_large_instance(domain, instance) do
     # servers may not publish peers
     peers =
@@ -94,7 +92,7 @@ defmodule Backend.Crawler.Crawlers.Mastodon do
         Map.take(instance, ["version", "description"]),
         Map.take(instance["stats"], ["user_count", "status_count"])
       )
-      |> Map.new(fn {k, v} -> {String.to_atom(k), v} end),
+      |> convert_keys_to_atoms(),
       %{
         peers: peers,
         interactions: interactions,