+++ backend/lib/backend/elasticsearch/cluster.ex
@@ -1,0 +1,3 @@
+defmodule Backend.Elasticsearch.Cluster do
+  use Elasticsearch.Cluster, otp_app: :backend
+end